feat(user_service): add endpoint et dto login +register

+++ user-service/src/main/java/com/fightclub/user_service/repositories/AppUserRepository.java
@@ -5,7 +5,7 @@ import com.fightclub.user_service.entities.UserEntity;
 
 import java.util.Optional;
 
-public interface AppUserRepository extends JpaRepository<UserEntity, Long> {
+public interface AppUserRepository extends JpaRepository<UserEntity, Integer> {
 
     UserEntity findUserEntityByPseudo(String pseudo);
